+++ include/power.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void resetActivityTimer();
+void enterDeepSleep();
+unsigned long getLastActivityTime();